ISS-53 # Loop enabled report -corrected - Useful for parallel run
