+++ src/test/resources/11_reports/02.1_loop_scenario_only.json
@@ -1,5 +1,5 @@
 {
-    "scenarioName": "Will Get A bath Room Multi Multi",
+    "scenarioName": "Loop Scenario - Will Get A bath Room",
     "loop": 2,
     "steps": [
         {
@@ -27,7 +27,6 @@
             "name": "get_another_shower_room",
             "url": "http://localhost:9998/home/bathroom/2${$.get_shower_room.response.body.id}",
             "operation": "GET",
-            "loop": 3,
             "request": {
                 "headers": {
                     "Content-Type": "application/json;charset=UTF-8",
