+++ src/test/resources/11_reports/03_multi_scenario_expected_report.json
@@ -1,43 +1,27 @@
 {
-  "results": [
-    {
-      "scenarioName": "Will Get A bath Room Multi Multi",
-      "steps": [
+    "results": [
         {
-          "name": "get_shower_room"
+            "scenarioName": "Loop Scenario - Will Get A bath Room",
+            "steps": [
+                {
+                    "name": "get_shower_room"
+                },
+                {
+                    "name": "get_another_shower_room"
+                }
+            ]
         },
         {
-          "name": "get_another_shower_room"
-        },
-        {
-          "loop": 1,
-          "name": "get_another_shower_room1"
-        },
-        {
-          "loop": 2,
-          "name": "get_another_shower_room2"
+            "scenarioName": "Loop Scenario - Will Get A bath Room",
+            "loop": 1,
+            "steps": [
+                {
+                    "name": "get_shower_room"
+                },
+                {
+                    "name": "get_another_shower_room"
+                }
+            ]
         }
-      ]
-    },
-    {
-      "scenarioName": "Will Get A bath Room Multi Multi",
-      "loop": 1,
-      "steps": [
-        {
-          "name": "get_shower_room"
-        },
-        {
-          "name": "get_another_shower_room"
-        },
-        {
-          "loop": 1,
-          "name": "get_another_shower_room1"
-        },
-        {
-          "loop": 2,
-          "name": "get_another_shower_room2"
-        }
-      ]
-    }
-  ]
+    ]
 }
